+++ src/background/processTextNodes.ts
@@ -1,5 +1,5 @@
 import { AppResponse } from '../common/types';
-import { PageMetadata, TextNode, WebpageDiacritizationData as WebpageDiacritizationData } from "../common/webpageDataClass";
+import { TextNode, WebpageDiacritizationData } from "../common/webpageDataClass";
 import { arabicToArabizi } from "./arabizi";
 import { fullDiacritization } from "./arabicDiacritization";
 import { messageContentScript, dataManager, controllerMap } from './background';
@@ -16,14 +16,15 @@ export async function processSelectedText(tab: chrome.tabs.Tab, method: string =
     return;
   }
 
-  const { selectedNodes } = request;
-  if (!selectedNodes) {
+  console.log(request);
+  const selectedNodes = new Set(request.selectedNodes);
+  if ( !selectedNodes ) {
     console.log("No text selected, processing full webpage.");
     await processWebpage(tab, method);
     return;
   }
 
-  console.log(`Processing ${method} for:`, selectedNodes);
+  console.log(`Processing ${method} for:`, selectedNodes.size, selectedNodes);
   await fullDiacritization(tab.id, tab.url, selectedNodes, controller.signal, method === 'arabizi')
 }
 
@@ -35,99 +36,41 @@ export async function processWebpage(tab: chrome.tabs.Tab, method: string): Prom
   const controller = new AbortController();
   controllerMap.set(tab.id, controller);
 
-  const response = await messageContentScript(tab.id, { action: 'getWebsiteMetadata' });
-  if (response.status === 'error') {
-    return response;
-  }
-  const { pageMetadata }: { pageMetadata?: PageMetadata; } = response;
-  if (!pageMetadata) {
-    return ({ status: 'error', error: new Error('Did not recieve pageMetadata from content') });
-  }
+  const latest = await messageContentScript(tab.id, { action: 'getWebsiteText' });
+  if (latest.status === 'error') return latest;
+  const { pageMetadata } = latest;
+  const selectedNodes = new Set<TextNode>(latest.selectedNodes);
+  if (!pageMetadata) throw new Error('No page metadata found');
+  if (!selectedNodes) throw new Error('No selected nodes found');
   const webpageDiacritizationData = await WebpageDiacritizationData.build(pageMetadata);
-  console.log('Website metadata:', pageMetadata);
+  await webpageDiacritizationData.createOriginal(selectedNodes);
 
-  // Load the saved data for the current webpage
-  const retrievedPageData = await dataManager.getWebpageData(tab.url);
-  console.log('Retrieved page data:', retrievedPageData);
-
-  let userMessage: string = '';
-
-  const requestMatchesSaved = function (): boolean {
-    // If there's no saved data, continue
-    if (!retrievedPageData) {
-      userMessage = "There's no saved webpage data, continuing";
-      return false;
-    }
-
-    // If saved data is available but has changed, ignore old data, continue
-    const oldSig = retrievedPageData.metadata.contentSignature;
-    const currentSig = pageMetadata.contentSignature;
-    if (oldSig !== currentSig) {
-      userMessage = `Content has changed, will regenerate diacritization and discard all old ones. Current hash: ${currentSig}, Saved hash: ${oldSig}`;
-      return false;
-      // TODO: only update the elements that have changed
-      // logChanges(retrievedPageData.metadata, pageMetadata);
-    }
-
-    // If saved data doesn't contain the requested diacritization method, retrieve all other saved diacritizations, continue
-    if (!Object.hasOwn(retrievedPageData.diacritizations, method)) {
-      webpageDiacritizationData.diacritizations = retrievedPageData.diacritizations;
-      userMessage = `Webpage is unchanged, generating ${method} from saved data: ${retrievedPageData.diacritizations}`;
-      return false;
-    }
-
-    // If saved data contains the requested method, update the website text with the saved data, stop
-    const diacritization = retrievedPageData.getDiacritization(method);
-    messageContentScript(tabId, { action: 'updateWebsiteText', tabUrl: tab.url, replacements: diacritization, method });
-    userMessage = `Webpage is unchanged, using saved ${method} data`;
-    return true;
-  };
-
-  // Stop here if we've already processed the request.
-  if (requestMatchesSaved()) {
-    return ({ status: 'success', userMessage });
-  }
-
-  // Get the website text
-  if (!webpageDiacritizationData.diacritizations['original']) {
-    const response = await messageContentScript(tab.id, { action: 'getWebsiteText' });
-    if (response.status === 'error') {
-      return response;
-    }
-    const { selectedNodes } = response;
-    if (!selectedNodes) {
-      return ({ status: 'error', error: new Error('Did not recieve TextNodes array from content script.') });
-    }
-    console.log('Retrieved original website text:', selectedNodes);
-    await webpageDiacritizationData.createOriginal(selectedNodes);
-  }
+  checkSaves();
 
   // Process the webpage
   console.log('Processing webpage:', tabUrl, 'with method:', method)
   switch (method) {
-    // If the method is 'fullDiacritics' and saved data exists for the current webpage, return the saved results
+    case 'original': {
+      const original = webpageDiacritizationData.getDiacritization('original');
+      messageContentScript(tabId, { action: 'updateWebsiteText', tabUrl, replacements: original, method });
+      break;
+    }
     case 'fullDiacritics':
       await fullDiacritization(tabId, tabUrl, webpageDiacritizationData.getDiacritization('original'), controller.signal)
-        .then((fullDiacritics) => {
-          webpageDiacritizationData.addDiacritization(fullDiacritics, method);
+        .then((result) => {
+          webpageDiacritizationData.addDiacritization(result, method);
         });
       break;
-
     case 'arabizi': {
       if (!webpageDiacritizationData.diacritizations['fullDiacritics']) {
         console.log("Full diacritization doesn't exist, Diacritizing text first");
         await fullDiacritization(tabId, tabUrl, webpageDiacritizationData.getDiacritization('original'), controller.signal, true)
-          .then((fullDiacritics) => {
-            webpageDiacritizationData.addDiacritization(fullDiacritics, 'fullDiacritics');
+          .then((result) => {
+            webpageDiacritizationData.addDiacritization(result, 'fullDiacritics');
           });
       }
-
-      const arabiziNodes: TextNode[] = arabicToArabizi(webpageDiacritizationData.getDiacritization('fullDiacritics'));
+      const arabiziNodes: Set<TextNode> = arabicToArabizi(webpageDiacritizationData.getDiacritization('fullDiacritics'));
       webpageDiacritizationData.addDiacritization(arabiziNodes, method);
-      break;
-    }
-
-    case 'original': {
       break;
     }
 
@@ -140,7 +83,7 @@ export async function processWebpage(tab: chrome.tabs.Tab, method: string): Prom
   dataManager.updateWebpageData(tab.url, webpageDiacritizationData)
     .then(() => {
       console.log('Saved webpage data updated:', webpageDiacritizationData)
-      message = { status: 'success', userMessage };
+      message = { status: 'success', userMessage: 'Webpage diacritization complete.' };
     })
     .catch((error) => {
       console.error('Failed to update saved webpage data:', error)
@@ -150,4 +93,48 @@ export async function processWebpage(tab: chrome.tabs.Tab, method: string): Prom
   messageContentScript(tabId, { action: 'allDone' });
   return message;
 
+}
+
+function checkSaves() {
+  // // Load the saved data for the current webpage
+  // const retrievedPageData = await dataManager.getWebpageData(tab.url);
+  // console.log('Retrieved page data:', retrievedPageData);
+
+  // let userMessage: string = '';
+
+  // const requestMatchesSaved = function (): boolean {
+  //   // If there's no saved data, continue
+  //   if (!retrievedPageData) {
+  //     userMessage = "There's no saved webpage data, continuing";
+  //     return false;
+  //   }
+
+  //   // If saved data is available but has changed, ignore old data, continue
+  //   const oldSig = retrievedPageData.metadata.contentSignature;
+  //   const currentSig = pageMetadata.contentSignature;
+  //   if (oldSig !== currentSig) {
+  //     userMessage = `Content has changed, will regenerate diacritization and discard all old ones. Current hash: ${currentSig}, Saved hash: ${oldSig}`;
+  //     return false;
+  //     // TODO: only update the elements that have changed
+  //     // logChanges(retrievedPageData.metadata, pageMetadata);
+  //   }
+
+  //   // If saved data doesn't contain the requested diacritization method, retrieve all other saved diacritizations, continue
+  //   if (!Object.hasOwn(retrievedPageData.diacritizations, method)) {
+  //     webpageDiacritizationData.diacritizations = retrievedPageData.diacritizations;
+  //     userMessage = `Webpage is unchanged, generating ${method} from saved data: ${retrievedPageData.diacritizations}`;
+  //     return false;
+  //   }
+
+  //   // If saved data contains the requested method, update the website text with the saved data, stop
+  //   const diacritization = retrievedPageData.getDiacritization(method);
+  //   messageContentScript(tabId, { action: 'updateWebsiteText', tabUrl: tab.url, replacements: diacritization, method });
+  //   userMessage = `Webpage is unchanged, using saved ${method} data`;
+  //   return true;
+  // };
+
+  // // Stop here if we've already processed the request.
+  // if (requestMatchesSaved()) {
+  //   return ({ status: 'success', userMessage });
+  // }
 }